+++ src/fr.inria.soctrace.framesoc.ui.gantt/src/fr/inria/soctrace/framesoc/ui/gantt/provider/GanttPresentationProvider.java
@@ -90,7 +90,7 @@ public class GanttPresentationProvider extends TimeGraphPresentationProvider {
     }
 
     @Override
-    public long getStateTableIndex(ITimeEvent event) {
+    public int getStateTableIndex(ITimeEvent event) {
         if (event instanceof TimeEvent && ((TimeEvent) event).hasValue()) {
             int type = ((TimeEvent) event).getValue();
             if (filteredTypes.contains(type)) {
@@ -106,7 +106,7 @@ public class GanttPresentationProvider extends TimeGraphPresentationProvider {
 	@Override
 	public String getEventName(ITimeEvent event) {
 
-		long index = getStateTableIndex(event);
+		int index = getStateTableIndex(event);
 		if (index >= 0 && index < stateTable.length) {
 			return stateTable[index].getStateString();
 		}
